- Added ADDTOCALLSTACK_INTENSIVE macro. On Windows builds this simply calls ADDTOCALLSTACK and on Linux builds this is ignored. This can be used in methods that are called very often in an attempt to alleviate the high CPU usage observed with the Linux build when ADDTOCALLSTACK is enabled (for some reason, the Linux build has a harder time with ADDTOCALLSTACK than Windows). Modified the following methods to use this: CSector::IsSectorSleeping, CSector::OnTick, CWorld::OnTick, and CWorld::GetSector. (Note that using this will reduce the accuracy of the call stacks on the Linux build, but it's better to at least have an estimated call stack than none at all) - Moved CSectorBase::GetIndex and GetMap to header file and removed ADDTOCALLSTACK from them. - Fixed the ASSERT in CServerDef::StatGet checking the index boundaries incorrectly. - Fixed a minor GCC warning in NetworkManager::acceptNewConnection.

+++ 0.56.b/network/network.cpp
@@ -2503,7 +2503,7 @@ void NetworkManager::acceptNewConnection(void)
 		else if ( climaxIp && ip.m_connected > climaxIp )
 			g_Log.Event(LOGM_CLIENTS_LOG|LOGL_ERROR, "Connection from %s rejected. (CLIENTMAXIP reached %ld/%ld)\n", (LPCTSTR)client_addr.GetAddrStr(), ip.m_connected, climaxIp);
 		else if ( ip.m_pings >= NETHISTORY_MAXPINGS )
-			g_Log.Event(LOGM_CLIENTS_LOG|LOGL_ERROR, "Connection from %s rejected. (MAXPINGS reached %ld/%ld)\n", (LPCTSTR)client_addr.GetAddrStr(), ip.m_pings, NETHISTORY_MAXPINGS );
+			g_Log.Event(LOGM_CLIENTS_LOG|LOGL_ERROR, "Connection from %s rejected. (MAXPINGS reached %ld/%ld)\n", (LPCTSTR)client_addr.GetAddrStr(), ip.m_pings, static_cast<long>(NETHISTORY_MAXPINGS) );
 		else
 			g_Log.Event(LOGM_CLIENTS_LOG|LOGL_ERROR, "Connection from %s rejected.\n", (LPCTSTR)client_addr.GetAddrStr());
